feat: don't need list_activities
Did some IA/wireframing (like I should have from the start) and figured
out that list_activities is redundant. Removing that method and
template.

Moving the list activities funcationality to the "show_city"
method/template.

Updating the url_for's on the Cancel buttons to go to the show_city
page.