del: don't need a special page for logged_in
Originally, I thought I needed an intermediate page for users logged-in
but not the creator of a city. That way the user can add activities to
another city (but they wouldn't be able to edit or delete the city
itself).

However, if the city creator tried to delete the city later, then he/she
would then be deleting activities made by another user.

The solution here is to not allow anyone but the city creator to add
activities to a city.