feat: edit fields only edit specified fields
Before, if you edited an item, you had to always fill out every field,
otherwise it would get cleared out.

Now the user has an option. They can click the checkbox, and that would
keep the existing text in the item, even if it was blank in the form and
then only change the fields that had non-empty-string values. Or, the
user can uncheck the box and clear out fields that they leave blank.

There's also a bug fix here. There was extraneous spaces in the
textarea, and apparently it was because if you leave space between the
opening and closing tags, it includes that space in the textarea value.
Now the opening and closing tags are next to each other.