fix: unrecognized URLs now go to 404
unrecognized URLs were resulting in a stack trace. There was a error
saying it was because the code couldn't find the login method.

I realized this was because base.html, error_403, and error_404 needed
to know the absolute paths for the methods (main.login, not .login).
Fixed that, and now 404s appear appropriately.