content: updating html headers and cities
Updating headers such that they use design items more consistently with
the rest of the site

Updating the list_cities page to recognize when no one is logged in and
no cities are there for display. It then shows text, encouraging the
user to sign in to get started. If the user is logged in and there are
no cities, then the Add City button is there.