feat: nonce check to decorator, adding to edit/new
Moved the nonce check functionality to a decorator function (those
things are cool! I got to erase a lot repeated code).

The nonce check was only on deletes before and now it also works on
edit_city, edit_activity, new_city, and new_activity

Tested functionality:
- Positive test cases like normal
- checked no nonce logic by deleting the nonce at the beginning of the
POST logic in the decorator
- checked wrong nonce logic by changing the nonce at the beginning of
the POST logic in the decorator